chore: major refactoring of PDE_Integrator_Test, testing for coupled systems with elimitated Dirichlet conditions chore: PDE_Integrator is now templatised to work with both SparseMatrix and CompressedRowMatrix, choose SPM for flexibility! chore: generalisation of sparsity pattern generation for CompressedRowMatrix to poly-element meshes development: new operators for cross coupling terms chores: improved readability of error messages and general reporting debug: fixed more bugs in vsetMakers

+++ source/includes/math/SparseMatrix.h
@@ -56,7 +56,7 @@ class SparseMatrix {
     colsConstIterator   RowEnd( size_t i ) const;
 
     /// zeroing out rows in the context of parallel computations
-    void      RemoveHalo( int32_t nrhalo );
+    void      RemoveHalo( size_t nrhalo );
     bool      HasEntry( size_t i, size_t j ) const { return data_[i].find(j) == data_[i].end(); }
     void      RemoveEntry( size_t, size_t );
     void      Erase();
